requirements.txt
absl-py==1.4.0
accelerate @ git+https://github.com/huggingface/accelerate.git@39d255b3d00b0a78725bd92eab9cd71d72ccceff
aiohttp==3.8.4
aiosignal==1.3.1
ansiwrap==0.8.4
anyio==3.7.1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.2.1
async-lru==2.0.4
async-timeout==4.0.2
attributedict==0.3.0
attrs==23.1.0
autoawq==0.1.7
Babel==2.13.1
backcall==0.2.0
beautifulsoup4==4.12.2
bitsandbytes==0.41.2.post2
black==23.3.0
bleach==6.0.0
blessings==1.7
cachetools==5.3.1
certifi==2023.5.7
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.1.0
click==8.1.3
cmake==3.26.4
codecov==2.1.13
colorama==0.4.6
coloredlogs==15.0.1
colour-runner==0.1.1
comm==0.1.3
contourpy==1.1.0
coverage==7.3.2
cupy==12.1.0
cycler==0.11.0
DataProperty==1.0.1
debugpy==1.6.7
decorator==5.1.1
deepdiff==6.7.1
deepspeed==0.9.5
defusedxml==0.7.1
diffusers==0.18.2
distlib==0.3.7
distro==1.8.0
docstring-parser==0.15
einops==0.7.0
entrypoints==0.4
evaluate==0.4.0
exceptiongroup==1.1.3
executing==1.2.0
fastjsonschema==2.17.1
fastrlock==0.8.1
filelock==3.13.1
fonttools==4.40.0
fqdn==1.5.1
frozenlist==1.3.3
fsspec==2023.6.0
google-auth==2.21.0
google-auth-oauthlib==1.0.0
grpcio==1.56.0
h11==0.14.0
hjson==3.1.0
httpcore==1.0.2
httpx==0.25.1
huggingface-hub==0.17.3
humanfriendly==10.0
idna==3.4
igraph==0.11.3
importlib-metadata==6.7.0
importlib-resources==5.12.0
inspecta==0.1.3
ipykernel==6.23.3
ipython==8.14.0
ipywidgets==7.0.0
isoduration==20.11.0
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
json5==0.9.14
jsonlines==4.0.0
jsonpickle==3.0.2
jsonpointer==2.4
jsonschema==4.19.2
jsonschema-specifications==2023.7.1
jupyter-events==0.9.0
jupyter-lsp==2.2.0
jupyter_client==8.3.0
jupyter_core==5.3.1
jupyter_server==2.10.0
jupyter_server_terminals==0.4.4
jupyterlab==4.0.8
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.9
jupyterlab_server==2.25.0
kiwisolver==1.4.4
Levenshtein==0.23.0
lit==16.0.6
lm-eval==0.3.0
Markdown==3.4.3
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.7.1
matplotlib-inline==0.1.6
mbstrdecoder==1.1.3
mdurl==0.1.2
mistune==3.0.1
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
nbclient==0.8.0
nbconvert==7.6.0
nbformat==5.9.0
nest-asyncio==1.5.6
networkx==3.1
ninja==1.11.1
nltk==3.8.1
notebook==7.0.6
notebook_shim==0.2.3
numexpr==2.8.7
numpy==1.25.0
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-ml-py==11.525.112
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.2.2
openai==1.3.4
ordered-set==4.1.0
overrides==7.4.0
packaging==23.1
pandas==2.0.2
pandocfilters==1.5.0
papermill==2.4.0
parso==0.8.3
pathos==0.3.1
pathspec==0.11.1
pathvalidate==3.2.0
peft==0.5.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==4.0.0
pluggy==1.3.0
portalocker==2.8.2
pox==0.3.3
ppft==1.7.6.7
prometheus-client==0.18.0
prompt-toolkit==3.0.38
protobuf==4.23.3
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyarrow==12.0.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pybind11==2.11.1
pycountry==22.3.5
pycparser==2.21
pydantic==1.10.9
pydot==1.4.2
Pygments==2.15.1
pygraphviz==1.11
pylcs==0.1.1
pyparsing==3.1.0
pyproject-api==1.6.1
pyre-extensions==0.0.29
pyrsistent==0.19.3
pytablewriter==1.2.0
python-dateutil==2.8.2
python-json-logger==2.0.7
pytz==2023.3
pyvis==0.3.2
PyYAML==6.0
pyzmq==25.1.0
rapidfuzz==3.5.2
referencing==0.30.2
regex==2023.6.3
requests==2.31.0
requests-oauthlib==1.3.1
responses==0.18.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.6.0
rootpath==0.1.1
rouge-score==0.1.2
rpds-py==0.12.0
rsa==4.9
sacrebleu==1.5.0
safetensors==0.3.1
scikit-learn==1.3.2
scipy==1.11.0
seaborn==0.12.2
Send2Trash==1.8.2
sentence-transformers==2.2.2
sentencepiece==0.1.99
shtab==1.6.4
six==1.16.0
sniffio==1.3.0
soupsieve==2.4.1
sqlitedict==2.1.0
stack-data==0.6.2
suffix-trees==0.3.0
sympy==1.12
tabledata==1.3.3
tabulate==0.9.0
tcolorpy==0.1.4
tenacity==8.2.2
tensorboard==2.13.0
tensorboard-data-server==0.7.1
tensorboardX==2.6.1
termcolor==2.3.0
terminado==0.17.1
texttable==1.7.0
textwrap3==0.9.2
threadpoolctl==3.1.0
tinycss2==1.2.1
tokenizers==0.14.1
toml==0.10.2
tomli==2.0.1
torch==2.0.1
tornado==6.3.2
tox==4.11.3
tqdm==4.65.0
tqdm-multiprocess==0.0.11
traitlets==5.9.0
transformers==4.35.0
triton==2.0.0
trl==0.7.2
typepy==1.3.2
types-python-dateutil==2.8.19.14
typing-inspect==0.9.0
typing_extensions==4.6.3
tyro==0.5.10
tzdata==2023.3
uri-template==1.3.0
urllib3==1.26.16
virtualenv==20.24.7
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.4
Werkzeug==2.3.6
widgetsnbextension==3.0.8
xformers==0.0.20
xxhash==3.2.0
yarl==1.9.2
zipp==3.15.0
zstandard==0.22.0
